# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DipWise API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/prices/{symbol}": {
      "get": {
        "tags": [
          "prices"
        ],
        "summary": "Get Prices",
        "description": "Return historical daily prices for a symbol (DB-backed with provider backfill).",
        "operationId": "get_prices_prices__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End"
            }
          },
          {
            "name": "sma_periods",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "SMA lookbacks (e.g. 100). Values appear in sma_by_period and, for period 100 only, sma_100.",
              "title": "Sma Periods"
            },
            "description": "SMA lookbacks (e.g. 100). Values appear in sma_by_period and, for period 100 only, sma_100."
          },
          {
            "name": "ema_periods",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "EMA spans (e.g. 100). Values appear in ema_by_period and, for period 100 only, ema_100.",
              "title": "Ema Periods"
            },
            "description": "EMA spans (e.g. 100). Values appear in ema_by_period and, for period 100 only, ema_100."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prices/{symbol}/refresh": {
      "post": {
        "tags": [
          "prices"
        ],
        "summary": "Refresh Prices",
        "description": "Force re-fetch from market data provider for the date range.",
        "operationId": "refresh_prices_prices__symbol__refresh_post",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End"
            }
          },
          {
            "name": "sma_periods",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "SMA lookbacks; warms up history before start when set.",
              "title": "Sma Periods"
            },
            "description": "SMA lookbacks; warms up history before start when set."
          },
          {
            "name": "ema_periods",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "EMA spans; combined with sma_periods for warmup window.",
              "title": "Ema Periods"
            },
            "description": "EMA spans; combined with sma_periods for warmup window."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/saved-backtests/": {
      "get": {
        "tags": [
          "saved-backtests"
        ],
        "summary": "List Saved Backtests Route",
        "operationId": "list_saved_backtests_route_saved_backtests__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedBacktestResponse"
                  },
                  "type": "array",
                  "title": "Response List Saved Backtests Route Saved Backtests  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "saved-backtests"
        ],
        "summary": "Create Saved Backtest Route",
        "operationId": "create_saved_backtest_route_saved_backtests__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedBacktestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedBacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/saved-backtests/{backtest_id}": {
      "get": {
        "tags": [
          "saved-backtests"
        ],
        "summary": "Get Saved Backtest Route",
        "operationId": "get_saved_backtest_route_saved_backtests__backtest_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedBacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PriceBar": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "adj_close": {
            "type": "number",
            "title": "Adj Close"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          },
          "sma_100": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sma 100",
            "description": "100-day simple moving average of adj_close when requested via sma_periods."
          },
          "ema_100": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ema 100",
            "description": "100-day exponential moving average (EWMA span) when requested via ema_periods."
          },
          "sma_by_period": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Sma By Period",
            "description": "SMA values keyed by period string (e.g. '50') for each requested sma_period."
          },
          "ema_by_period": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Ema By Period",
            "description": "EMA values keyed by period string for each requested ema_period."
          }
        },
        "type": "object",
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close",
          "adj_close",
          "volume"
        ],
        "title": "PriceBar"
      },
      "PriceListResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PriceBar"
            },
            "type": "array",
            "title": "Prices"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "start",
          "end",
          "prices"
        ],
        "title": "PriceListResponse"
      },
      "SavedBacktestCreate": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "parameters",
          "results"
        ],
        "title": "SavedBacktestCreate"
      },
      "SavedBacktestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Results"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "parameters",
          "results",
          "created_at"
        ],
        "title": "SavedBacktestResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from POST /auth/login"
      }
    }
  }
}